% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pushshift_io.R
\name{get_comment_reddit}
\alias{get_comment_reddit}
\title{Get comments from pushshift.io}
\usage{
get_comment_reddit(subreddit = "all", author = NULL, n = 1000,
  after = NULL)
}
\arguments{
\item{subreddit}{Name of subreddit from which to get data. Defaults to "all".}

\item{author}{Restrict results to author.}

\item{n}{Number of submission/posts to return. Defaults to 1000.}

\item{after}{Optional, the date-time from which to start the next search.}

\item{before}{Optional, the date-time from which to start the next search.}
}
\value{
A data frame of reddit data.
}
\description{
Reads/parses reddit data from api.pushshift.io
}
\details{
Column descriptions are provided below

\itemize{
  \item author Name of author of post
  \item author_flair_css_class Flair class attribute
  \item author_flair_text Flair text
  \item body Text of the comment
  \item created_utc Date-time in UTC timezone
  \item domain Domain
  \item full_link Full URL
  \item id Unique identifier string
  \item is_self Whether is self
  \item link_flair_css_class Flair class attribute for link
  \item link_flair_text Flair text for link
  \item num_comments Number of comments
  \item over_18 Whether material is intended for people over 18
  \item permalink Stable link
  \item post_hint Type of post
  \item retrieved_on Date-time when data was originally received
  \item subreddit Name of subreddit
  \item score Reddit score
  \item selftext Text
  \item stickied Whether it's been stickied
  \item subreddit Name of subreddit
  \item subreddit_id Unique identifier of subreddit
  \item thumbnail Path to post thumbnail
  \item title Title of post
  \item url URL
}
}
