% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reddit_com.R
\name{get_reddit_com}
\alias{get_reddit_com}
\title{Get data from reddit.com}
\usage{
get_reddit_com(subreddit = "all", n = 100, after = NULL,
  before = NULL)
}
\arguments{
\item{subreddit}{Name of subreddit from which to get data. Defaults to "all".}

\item{n}{Number of submission/posts to return. Defaults to 100.}

\item{after}{Optional, the parameter from which to start the next search.}

\item{before}{Optional, the parameter from which to start the next search.}
}
\value{
A data frame of reddit data.
}
\description{
Reads/parses data directly from reddit.com
}
